+++ src/listpack.h
@@ -55,14 +55,14 @@ unsigned char *lpPushHead(unsigned char *lp, unsigned char *s, uint32_t slen);
 unsigned char *lpPushTail(unsigned char *lp, unsigned char *s, uint32_t slen);
 unsigned char *lpReplace(unsigned char *lp, unsigned char *p, unsigned char *s, uint32_t slen);
 unsigned char *lpDelete(unsigned char *lp, unsigned char *p, unsigned char **newp);
-uint32_t lpLength(unsigned char *lp);
+long lpLength(unsigned char *lp);
 unsigned char *lpGet(unsigned char *p, int64_t *count, unsigned char *intbuf);
 unsigned char *lpFind(unsigned char *lp, unsigned char *p, unsigned char *s, unsigned int slen, unsigned int skip);
 unsigned char *lpFirst(unsigned char *lp);
 unsigned char *lpLast(unsigned char *lp);
 unsigned char *lpNext(unsigned char *lp, unsigned char *p);
 unsigned char *lpPrev(unsigned char *lp, unsigned char *p);
-uint32_t lpBytes(unsigned char *lp);
+size_t lpBytes(unsigned char *lp);
 unsigned char *lpSeek(unsigned char *lp, long index);
 typedef int (*listpackValidateEntryCB)(unsigned char *p, void *userdata);
 int lpValidateIntegrity(unsigned char *lp, size_t size, int deep,
